Fikser interne lenker i LinkPanels
