+++ src/components/_common/lenkepanel-liste/LenkepanelListe.less
@@ -1,4 +1,4 @@
-@import '../../../../global';
+@import '../../../global';
 
 .lenkepanel-liste {
     width: 100%;
