+++ src/components/parts/link-panels/LinkPanels.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { LenkepanelListe } from './lenkepanel-liste/LenkepanelListe';
+import { LenkepanelListe } from '../../_common/lenkepanel-liste/LenkepanelListe';
 import { ContentProps } from 'types/content-props/_content-common';
 import './LinkPanels.less';
